test (webapi): fix integration and sim tests
