+++ webapi/src/it/scala/org/knora/webapi/ITKnoraLiveSpec.scala
@@ -19,15 +19,23 @@
 
 package org.knora.webapi
 
+import java.io.File
+import java.nio.file.{Files, Paths}
+
 import akka.actor.ActorSystem
 import akka.event.LoggingAdapter
+import akka.http.scaladsl.Http
+import akka.http.scaladsl.client.RequestBuilding
+import akka.http.scaladsl.model.{HttpRequest, HttpResponse, StatusCodes}
 import akka.stream.ActorMaterializer
 import com.typesafe.config.{Config, ConfigFactory}
 import org.knora.webapi.messages.app.appmessages.SetAllowReloadOverHTTPState
 import org.knora.webapi.util.StringFormatter
-import org.scalatest.{BeforeAndAfterAll, Suite}
+import org.scalatest.{BeforeAndAfterAll, Matchers, Suite, WordSpecLike}
+import spray.json.{JsObject, _}
 
-import scala.concurrent.ExecutionContext
+import scala.concurrent.duration.{Duration, _}
+import scala.concurrent.{Await, ExecutionContext, Future}
 import scala.languageFeature.postfixOps
 
 object ITKnoraLiveSpec {
@@ -38,14 +46,14 @@ object ITKnoraLiveSpec {
   * This class can be used in End-to-End testing. It starts the Knora server and
   * provides access to settings and logging.
   */
-class ITKnoraLiveSpec(_system: ActorSystem) extends Core with KnoraService with Suite with BeforeAndAfterAll {
-
-    /* needed by the core trait */
-    implicit lazy val settings: SettingsImpl = Settings(system)
+class ITKnoraLiveSpec(_system: ActorSystem) extends Core with KnoraService with Suite with WordSpecLike with Matchers with BeforeAndAfterAll with RequestBuilding  {
 
     implicit val materializer: ActorMaterializer = ActorMaterializer()
 
     implicit val executionContext: ExecutionContext = system.dispatchers.defaultGlobalDispatcher
+
+    /* needed by the core trait */
+    implicit lazy val settings: SettingsImpl = Settings(system)
 
     StringFormatter.initForTest()
 
@@ -80,6 +88,46 @@ class ITKnoraLiveSpec(_system: ActorSystem) extends Core with KnoraService with 
         /* Stop the server when everything else has finished */
         log.debug(s"Stopping Knora Service")
         stopService()
+    }
+
+    protected def singleAwaitingRequest(request: HttpRequest, duration: Duration = 15.seconds): HttpResponse = {
+        val responseFuture = Http().singleRequest(request)
+        Await.result(responseFuture, duration)
+    }
+
+    protected def getResponseString(request: HttpRequest): String = {
+        val response = singleAwaitingRequest(request)
+
+        //log.debug("REQUEST: {}", request)
+        //log.debug("RESPONSE: {}", response.toString())
+
+        val responseBodyFuture: Future[String] = response.entity.toStrict(5.seconds).map(_.data.decodeString("UTF-8"))
+        val responseBodyStr = Await.result(responseBodyFuture, 5.seconds)
+
+        assert(response.status === StatusCodes.OK, s",\n REQUEST: $request,\n RESPONSE: $response")
+        responseBodyStr
+    }
+
+    protected def checkResponseOK(request: HttpRequest): Unit = {
+        getResponseString(request)
+    }
+
+    protected def getResponseJson(request: HttpRequest): JsObject = {
+        getResponseString(request).parseJson.asJsObject
+    }
+
+    /**
+      * Creates the Knora API server's temporary upload directory if it doesn't exist.
+      */
+    def createTmpFileDir(): Unit = {
+        if (!Files.exists(Paths.get(settings.tmpDataDir))) {
+            try {
+                val tmpDir = new File(settings.tmpDataDir)
+                tmpDir.mkdir()
+            } catch {
+                case e: Throwable => throw FileWriteException(s"Tmp data directory ${settings.tmpDataDir} could not be created: ${e.getMessage}")
+            }
+        }
     }
 
 }
