--- webapi/src/it/scala/org/knora/webapi/util/TestingUtilities.scala
@@ -1,82 +1,0 @@
-/*
- * Copyright © 2015-2018 the contributors (see Contributors.md).
- *
- * This file is part of Knora.
- *
- * Knora is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published
- * by the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * Knora is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public
- * License along with Knora.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-package org.knora.webapi.util
-
-import java.io.File
-import java.nio.file.{Files, Paths}
-
-import akka.http.scaladsl.Http
-import akka.http.scaladsl.client.RequestBuilding
-import akka.http.scaladsl.model.{HttpRequest, HttpResponse, StatusCodes}
-import org.knora.webapi.{Core, FileWriteException}
-import org.scalatest.{BeforeAndAfterAll, Matchers, WordSpecLike}
-import spray.json.{JsObject, _}
-
-import scala.concurrent.duration._
-import scala.concurrent.{Await, Future}
-
-/**
-  * Created by subotic on 26.06.17.
-  */
-trait TestingUtilities extends WordSpecLike with Matchers with BeforeAndAfterAll with RequestBuilding {
-
-    this: Core =>
-
-    protected def singleAwaitingRequest(request: HttpRequest, duration: Duration = 15.seconds): HttpResponse = {
-        val responseFuture = Http().singleRequest(request)
-        Await.result(responseFuture, duration)
-    }
-
-    protected def getResponseString(request: HttpRequest): String = {
-        val response = singleAwaitingRequest(request)
-
-        //log.debug("REQUEST: {}", request)
-        //log.debug("RESPONSE: {}", response.toString())
-
-        val responseBodyFuture: Future[String] = response.entity.toStrict(5.seconds).map(_.data.decodeString("UTF-8"))
-        val responseBodyStr = Await.result(responseBodyFuture, 5.seconds)
-
-        assert(response.status === StatusCodes.OK, s",\n REQUEST: $request,\n RESPONSE: $response")
-        responseBodyStr
-    }
-
-    protected def checkResponseOK(request: HttpRequest): Unit = {
-        getResponseString(request)
-    }
-
-    protected def getResponseJson(request: HttpRequest): JsObject = {
-        getResponseString(request).parseJson.asJsObject
-    }
-
-    /**
-      * Creates the Knora API server's temporary upload directory if it doesn't exist.
-      */
-    def createTmpFileDir(): Unit = {
-        if (!Files.exists(Paths.get(settings.tmpDataDir))) {
-            try {
-                val tmpDir = new File(settings.tmpDataDir)
-                tmpDir.mkdir()
-            } catch {
-                case e: Throwable => throw FileWriteException(s"Tmp data directory ${settings.tmpDataDir} could not be created: ${e.getMessage}")
-            }
-        }
-    }
-
-}
