+++ webapi/src/test/resources/logback-test.xml
@@ -59,14 +59,14 @@
     <logger name="org.knora" level="INFO"/>
     <logger name="org.knora.webapi" level="INFO"/>
     <logger name="org.knora.webapi.KnoraExceptionHandler" level="INFO"/>
-    <logger name="org.knora.webapi.app.ApplicationStateActor" level="DEBUG"/>
+    <logger name="org.knora.webapi.app.ApplicationStateActor" level="INFO"/>
     <logger name="org.knora.webapi.util.cache" level="INFO"/>
     <logger name="org.knora.webapi.util.PermissionUtilADM" level="INFO"/>
     <logger name="org.knora.webapi.util.TransformData" level="INFO"/>
     <logger name="org.knora.webapi.routing.Authenticator" level="INFO"/>
     <logger name="org.knora.webapi.routing.JWTHelper" level="INFO"/>
     <logger name="org.knora.webapi.routing.HealthRoute" level="INFO"/>
-    <logger name="org.knora.webapi.routing.RejectingRoute" level="DEBUG"/>
+    <logger name="org.knora.webapi.routing.RejectingRoute" level="INFO"/>
     <logger name="org.knora.webapi.routing.v1.AssetsRouteV1" level="INFO"/>
     <logger name="org.knora.webapi.routing.v1.ResourcesRouteV1" level="INFO"/>
     <logger name="org.knora.webapi.routing.v1.StandoffRouteV1" level="INFO"/>
@@ -121,7 +121,7 @@
     <logger name="org.knora.webapi.e2e.HealthRouteE2ESpec" level="INFO"/>
     <logger name="org.knora.webapi.e2e.admin.ListsADME2ESpec" level="INFO"/>
     <logger name="org.knora.webapi.e2e.admin.ProjectsADME2ESpec" level="INFO"/>
-    <logger name="org.knora.webapi.e2e.admin.UsersADME2ESpec" level="DEBUG"/>
+    <logger name="org.knora.webapi.e2e.admin.UsersADME2ESpec" level="INFO"/>
 
 
     <root level="INFO">
